requirements.txt
Flask==1.0.2
Flask-Sqlalchemy==2.3.2
flask-migrate==2.2.1
flask-script==2.0.6